refactor(teams): split plugin module into submodules [OPE-368]
